tools/Space-Ninja-Web/requirements.txt
Flask
python-pptx
Werkzeug
